+++ tw-tasks-executor/src/main/java/com/transferwise/MyApp.java
@@ -1,0 +1,8 @@
+package com.transferwise;
+
+public class MyApp {
+
+  public static void main(String[] args) {
+    System.out.println(Runtime.getRuntime().availableProcessors());
+  }
+}