+++ tw-tasks-executor/src/main/java/com/transferwise/tasks/handler/ExponentialTaskRetryPolicy.java
@@ -1,5 +1,6 @@
 package com.transferwise.tasks.handler;
 
+import com.google.common.math.LongMath;
 import com.transferwise.common.baseutils.clock.ClockHolder;
 import com.transferwise.tasks.domain.ITask;
 import com.transferwise.tasks.handler.interfaces.ITaskRetryPolicy;
@@ -7,6 +8,7 @@ import java.time.Duration;
 import java.time.ZonedDateTime;
 import java.time.temporal.ChronoUnit;
 import lombok.Getter;
+import lombok.NonNull;
 import lombok.Setter;
 import lombok.experimental.Accessors;
 
@@ -15,14 +17,16 @@ import lombok.experimental.Accessors;
 @Accessors(chain = true)
 public class ExponentialTaskRetryPolicy implements ITaskRetryPolicy {
 
+  @NonNull
   private Duration delay = Duration.ofMillis(0);
-  private double multiplier = 2;
+  private long multiplier = 2;
   /**
    * Maximum delay to add on each retry (not the total delay since task creation).
    */
-  private Duration maxDelay;
-  private int maxCount = 1;
-  private int triesDelta = 0;
+  @NonNull
+  private Duration maxDelay = Duration.ofDays(1);
+  private long maxCount = 1;
+  private long triesDelta = 0;
 
   @Override
   public ZonedDateTime getRetryTime(ITask taskForProcessing, Throwable t) {
@@ -30,10 +34,21 @@ public class ExponentialTaskRetryPolicy implements ITaskRetryPolicy {
     if (triesCount >= maxCount) {
       return null;
     }
-    long addedTimeMs = delay.toMillis() * (long) Math.pow(multiplier, (triesCount - 1));
-    if (maxDelay != null && addedTimeMs > maxDelay.toMillis()) {
+    long addedTimeMs;
+
+    try {
+      if (triesCount > Integer.MAX_VALUE) {
+        addedTimeMs = maxDelay.toMillis();
+      } else {
+        addedTimeMs = LongMath.checkedMultiply(delay.toMillis(), LongMath.checkedPow(multiplier, (int) triesCount - 1));
+        if (addedTimeMs > maxDelay.toMillis()) {
+          addedTimeMs = maxDelay.toMillis();
+        }
+      }
+    } catch (ArithmeticException ignored) {
       addedTimeMs = maxDelay.toMillis();
     }
+
     return ZonedDateTime.now(ClockHolder.getClock()).plus(addedTimeMs, ChronoUnit.MILLIS);
   }
 }